+++ src/main/java/step4/Application.java
@@ -4,7 +4,7 @@ import step4.util.InputUtils;
 import step4.vehicle.Cars;
 import step4.view.ResultView;
 
-public class Start {
+public class Application {
     public static void main(String[] args) {
         Cars cars = new Cars(InputUtils.inputCarNameView());
         ResultView resultView = new ResultView();
