+++ src/main/java/step4/vehicle/Cars.java
@@ -2,13 +2,16 @@ package step4.vehicle;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class Cars {
+    public static final int NUMBER_OF_PLAYERS = 2;
+
     private final List<Car> cars = new ArrayList<>();
     private final List<Car> winners = new ArrayList<>();
 
     public Cars(String[] names) {
-        if (names.length < 2) {
+        if (names.length < NUMBER_OF_PLAYERS) {
             throw new IllegalArgumentException("게임을 진행할 수 없음");
         }
         initCars(names);
@@ -24,22 +27,11 @@ public class Cars {
         return cars;
     }
 
-    public List<Car> getWinners() {
+    public List<Car> findWinners() {
         int maxScore = cars.stream().mapToInt(Car::getLocation).max().getAsInt();
-        findWinner(maxScore);
-        return winners;
-    }
-
-    private void findWinner(int checkMax) {
-        for (Car car : cars) {
-            addWinner(checkMax, car);
-        }
-    }
-
-    private void addWinner(int checkMax, Car car) {
-        if (car.getLocation() == checkMax) {
-            winners.add(car);
-        }
+        return cars.stream()
+                .filter(car -> car.getLocation() == maxScore)
+                .collect(Collectors.toList());
     }
 
 }
